+++ src/components/shop/Shop.tsx
@@ -1,33 +1,6 @@
+import { items, mainItem } from '@/mocks/home/ShopItem';
+
 const Shop = () => {
-  const mainItem = {
-    src: 'https://images.pexels.com/photos/1480965/pexels-photo-1480965.jpeg?auto=compress&cs=tinysrgb&w=600',
-    title: 'Item 1',
-    price: '$ price',
-  };
-
-  const items = [
-    {
-      src: 'https://images.pexels.com/photos/843341/pexels-photo-843341.jpeg?auto=compress&cs=tinysrgb&w=600',
-      title: 'Item 2',
-      price: 'price',
-    },
-    {
-      src: 'https://images.pexels.com/photos/15442329/pexels-photo-15442329.jpeg?auto=compress&cs=tinysrgb&w=600&lazy=load',
-      title: 'Item 3',
-      price: 'price',
-    },
-    {
-      src: 'https://images.pexels.com/photos/264279/pexels-photo-264279.jpeg?auto=compress&cs=tinysrgb&w=600',
-      title: 'Item 4',
-      price: 'price',
-    },
-    {
-      src: 'https://images.pexels.com/photos/257970/pexels-photo-257970.jpeg?auto=compress&cs=tinysrgb&w=600',
-      title: 'Item 5',
-      price: 'price',
-    },
-  ];
-
   return (
     <>
       <div className="h-screen w-full bg-black px-48 py-16 text-white">
